Hawk: Move derived attribute computation into 'Initial' time
The derived attribute was being defined before the folder
was added, meaning that the precomputation of the scores in
Hawk and HawkIncUpdate solutions was being done in the 'Load'
stage instead of the 'Initial' stage.

This commit moves that definition of the derived attributes
and the derivation of the initial set of scores to the 'Initial'
phase.