Hawk: solve rare race condition
Hawk does the indexing in a background thread, so manually
requesting a synchronisation of the index and waiting for its
completion can be tricky. Typically, we do this with semaphores
and graph change listeners. The old order was this:

1. The TTC solution thread performs an action that will trigger
   a synchronisation of the index by the background thread.

2. The TTC solution thread adds a graph listener that will
   have the Hawk indexing thread release a semaphore when the
   sychronisation ends.

3. The TTC solution thread then blocks while acquiring the
   semaphore.

4. The semaphore is released by the indexing thread (see step 2)
   and the TTC solution proceeds its execution.

This approach had a rare race condition: if the synchronisation
was already done before 2 finished adding the graph listener, the
semaphore would never be released and 3 would block indefinitely.
This only very rarely happens, but it can happen across many runs
as in our benchmarks.

The fix for this is to do step 2 before step 1, and then the
semaphore will always be released even if the synchronisation completes
before step 3, and step 3 will eventually complete.

The reason why this only typically happens in the batch scenarios
is somewhat unclear, as the race condition could have been happening
across all options. The code that was showing the race condition was
common to the batch, IU and IUQ solutions. It may be that the higher
resource demands of the batch solution make this race condition more
likely to happen.