Hawk: create/restore backup at the start
In the batch solution, Hawk tracks initial.xmi as it would in its
most common use case. This means that initial.xmi needs to be updated
and saved. If a run fails, the backup was restored, but this restoration
could fail as well if the benchmark script killed the process in the
middle of the restoration.

This commit changes Hawk so it will create a backup before the first use
of a certain initial.xmi file, and it will restore that backup at the
beginning of later runs. This should prevent the restoration from being
impacted by a timeout.